HOMME: replaced Derivative with ReferenceElement
ReferenceElement contains information about the (pseudo-spectral)
derivative matrix (what was called dvv) and the SEM mass matrix.
The latter, is an array stored in each element in F90 (and that's
how it was in C++ too), but it is constant on each element (since
it refers to the mass matrix on the reference element). Therefore,
it is more appropriate (and less confusing) to store it just once.

A ReferenceElement seems appropriate in this context, to store
informations that are independent of the particular element, and
only require information on the reference finite element.