HOMME: moved ElementsState.*pp inside preqx_kokkos/cxx subfolder
The state is specific of the model. This change makes the kokkos target
more similar to the f90 one, where the element_state.F90 is placed in
the model subfolder.

Also, the random_init_ut test now depends on preqx-specific sources. Hence,
we created a preqx_kokkos_ut folder, with some preqx-related unit tests.